requirements.txt
numpy==1.19.5
pyyaml
click
mlflow